+++ src/app/Modules/Inventory/Domain/Events/StockMovementCreated.php
@@ -16,7 +16,7 @@ class StockMovementCreated
         public readonly int $productVariantId,
         public readonly int $warehouseId,
         public readonly string $movementType,
-        public readonly int $quantity,
+        public readonly float $quantity,
         public readonly ?int $warehouseToId = null,
     ) {}
 }
