+++ src/app/Modules/Inventory/Infrastructure/Http/Requests/StoreAdjustmentRequest.php
@@ -19,7 +19,7 @@ class StoreAdjustmentRequest extends FormRequest
             'product_variant_id' => ['required', 'integer', 'exists:product_variants,id'],
             'warehouse_id' => ['required', 'integer', 'exists:warehouses,id'],
             'location_id'  => ['required', 'integer', 'exists:locations,id'],
-            'quantity'     => ['required', 'integer', 'not_in:0'],
+            'quantity'     => ['required', 'numeric', 'not_in:0'],
             'reason'       => ['required', 'string'],
         ];
     }
